Fixed bug with SetDenseTableMessage that caused payload length to be interpreted as a byte, resulting in sign errors. Expanded toString() methods.

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/routing/RouteTableMessage.java
@@ -37,7 +37,7 @@ public abstract class RouteTableMessage extends Message {
      * The remaining parameters are the same as in Message's constructor.
      */
     protected RouteTableMessage(byte ttl,
-                                byte length, 
+                                int length, 
                                 byte variant,
                                 byte tableTTL) {
         super(Message.F_ROUTE_TABLE_UPDATE, ttl, length);
